refactor confirm modals

+++ components/ui/country-doc-card-upload.js
@@ -10,15 +10,14 @@ import DocumentationService from 'services/documentationService';
 import modal from 'services/modal';
 
 // Components
+import { showConfirmModal } from 'components/ui/confirm-modal';
 import CountryDocModal from 'components/ui/country-doc-modal';
-import Spinner from 'components/ui/spinner';
 import useUser from 'hooks/use-user';
 
 const CountryDocCardUpload = (props) => {
   const { status, docType, id, onChange } = props;
   const intl = useIntl();
   const user = useUser();
-  const [deleteLoading, setDeleteLoading] = useState(false);
 
   const documentationService = useMemo(() => new DocumentationService({
     authorization: user.token
@@ -41,17 +40,27 @@ const CountryDocCardUpload = (props) => {
   const triggerDeleteFile = (e) => {
     e && e.preventDefault();
 
-    setDeleteLoading(true);
-
-    documentationService.deleteDocument(id, 'gov-documents')
-      .then(() => {
-        setDeleteLoading(false);
-        onChange && onChange();
-      })
-      .catch((err) => {
-        setDeleteLoading(false);
-        console.error(err);
-      });
+    showConfirmModal({
+      text: intl.formatMessage(
+        { id: 'delete.document.text', defaultMessage: 'Are you sure you want to delete document {document}?' }, { document: title }
+      ),
+      confirmText: intl.formatMessage({ id: 'delete', defaultMessage: 'Delete' }),
+      onConfirm: ({ onSuccess, onError } = {}) => {
+        documentationService.deleteDocument(id, 'gov-documents')
+          .then(() => {
+            modal.toggleModal(false);
+            onSuccess && onSuccess();
+            onChange && onChange();
+          })
+          .catch((err) => {
+            onError && onError(
+              intl.formatMessage({ id: 'document.delete.error', defaultMessage: 'An error occurred while deleting the document.' })
+            );
+            Sentry.captureException(err);
+            console.error(err);
+          });
+      }
+    })
   };
 
   const classNames = classnames({
@@ -71,7 +80,6 @@ const CountryDocCardUpload = (props) => {
           <li>
             <button onClick={triggerDeleteFile} className="c-button -small -primary">
               {intl.formatMessage({ id: 'delete' })}
-              <Spinner isLoading={deleteLoading} className="-tiny -transparent" />
             </button>
           </li>
         </ul>
